lars-internal {lars}	R Documentation

Internal lars functions

Description

Internal lars functions

Usage

betabreaker(object)
backsolvet(r,x,k = ncol(r))
cv.folds(n, folds = 10)
delcol(r, z, k = p)
downdateR (R, k = p)
error.bars(x, upper, lower, width = 0.02, ...)
nnls.lars(active, Sign, R, beta, Gram, eps = 1e-10, trace = FALSE, 
    use.Gram = TRUE)
plotCVLars(cv.lars.object, se = TRUE)
updateR(xnew, R = NULL, xold, eps = .Machine$double.eps, Gram = FALSE)

Details

These are not to be called by the user. betabreaker figures out if coefficients (other than lasso) pass through zero, since the L1 norm is discontinuous there, and this has an impact on predict/plot. Suggested by Yann-Ael Le Borgne. backsolvet is included to make the R code compatible with the Splus code, since backsolve in R has a transpose=TRUE option already.

Author(s)

Trevor Hastie

[Package lars version 1.2 Index]