summary.lars {lars}	R Documentation

Summary method for lars objects

Description

Produce an anova-type summary for a lars object.

Usage

## S3 method for class 'lars'
summary(object, sigma2=NULL, ...)

Arguments

object
lars object

sigma2
optional variance measure (for p>n)

...
Additional arguments for summary generic

Details

An anova summary is produced, with Df, RSS and Cp for each step. Df is tricky for some models, such as forward stagewise and stepwise, and is not likely to be accurate. When p>n, the user is responsible for supplying sigma2.

Value

An anova object is returned, with rownames the step number, and with components:

Df
Estimated degree of freedom

Rss
The Residual sum of Squares

Cp
The Cp statistic

Author(s)

Brad Efron and Trevor Hastie

References

Efron, Hastie, Johnstone and Tibshirani (2003) "Least Angle Regression" (with discussion) Annals of Statistics; see also http://www-stat.stanford.edu/~hastie/Papers/LARS/LeastAngle_2002.pdf. Hastie, Tibshirani and Friedman (2002) Elements of Statistical Learning, Springer, NY.

See Also

lars, and print, plot,and predict methods for lars, and cv.lars

Examples

data(diabetes)
attach(diabetes)
object <- lars(x,y)
summary(object)
detach(diabetes)

[Package lars version 1.2 Index]